requirements.txt
pypdf==3.8.1
pinecone-client==2.2.1
tiktoken==0.3.3
langchain
chainlit==0.4.0
google-search-results
oidc-client